Refactor frontend classes
